% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_ridgereg.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{coef S3 method}
\usage{
\method{coef}{ridgereg}(object, ...)
}
\arguments{
\item{object}{An object of class ridgereg}

\item{...}{other arguments}
}
\value{
coefficients of linear regression
}
\description{
S3 method for objects of class ridgereg providing the coefficients of the linear regression
}
