% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred.linreg}
\alias{pred.linreg}
\title{pred S3 method}
\usage{
\method{pred}{linreg}(object, data = NULL, ...)
}
\arguments{
\item{object}{An object of class linreg}

\item{data}{A dataframe for which the values should be predicted. If not provided, the training data for the object is used.}

\item{...}{other arguments}
}
\value{
vector containing fitted values
}
\description{
S3 method for objects of class linreg providing the fitted values of the linear regression
}
