% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ridgereg.R
\name{pred.ridgereg}
\alias{pred.ridgereg}
\title{pred S3 method}
\usage{
\method{pred}{ridgereg}(object)
}
\arguments{
\item{object}{An object of class ridgereg}
}
\value{
vector containing fitted values
}
\description{
S3 method for objects of class ridgereg providing the fitted values of the linear regression
}
