% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ridgereg.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{print S3 method}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{An object of class ridgereg}

\item{...}{other arguments}
}
\value{
nothing. Prints tha call and the coefficients of the ridgereg object.
}
\description{
S3 method for objects of class ridgereg printing the call and the coefficients of the linear model
}
