% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Reference Class representing a Ridge Linear regression using least square}
\arguments{
\item{formula}{a formula object.}

\item{data}{a data frame.}

\item{lambda}{a numeric number.}
}
\value{
an object of class `ridgereg`
}
\description{
ridgereg is used to fit linear models with a penalty lambda and
  using ordinary linear algebra to calculate the Regressions coefficients and
  fitted values
}

\examples{
# data(iris)
# ridgereg_mod <- ridgereg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris, lambda = 0.5)
# l$coef()

 
}
