% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{@title Reference Class representing a Ridge Linear regression using least square}
\value{
an object of class `ridgereg`
}
\description{
ridgereg is used to fit linear models with a penalty lambda and
  using ordinary linear algebra to calculate the Regressions coefficients and
  fitted values
}
\section{Fields}{

\describe{
\item{\code{formula}}{a formula object.}

\item{\code{data}}{a data frame.}

\item{\code{lambda}}{a numeric number.}
}}

\examples{
#' # data(iris)
# l <- ridgereg$new(formula=Petal.Length~Sepal.Width+Sepal.Length, data=iris, lambda = 0.5)
# l$coef()

}
