% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualization of mean departure delay by airport.}
\usage{
visualize_airport_delays()
}
\value{
A plot of mean departure delay by airport where the size of the 
circle increases depending on the mean departure delay.
}
\description{
This function will show a ggplot where the circles will increase in size if
the mean departure delay is bigger at a specific airport.
}
