% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{ridgereg}
\source{
http://staff.www.ltu.se/~jove/courses/c0002m/least_squares.pdf

http://pages.stat.wisc.edu/~larget/math496/qr.html

https://math.stackexchange.com/questions/299481/qr-factorization-for-ridge-regression
}
\usage{
ridgereg(formula, data, lambda)
}
\arguments{
\item{formula}{A formula stating the form of the regression}

\item{data}{A dataframe to use for the regression}

\item{lambda}{A numeric representing the lambda value for the ridge regression}
}
\value{
An object of class linreg containing regression coefficients, fitted values, residuals, degree-of-freedom, residual variance, variance of the regression coefficient, t-values and p-values
}
\description{
A function that performs a ridge regression based on QR decomposition
}
