% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{visualize_airport_delays}
\source{
https://jrnold.github.io/r4ds-exercise-solutions/relational-data.html
}
\usage{
visualize_airport_delays()
}
\value{
NA
}
\description{
A function visualizing the mean delay of flights for different airports
}
